+++ admin/settings.asset.php
@@ -1,0 +1,1 @@
+<?php return array( 'dependencies' => array( 'wp-element', 'wp-components', 'wp-api-fetch', 'wp-i18n', 'wp-dom-ready' ), 'version' => '0.2.0' );